+++ view/theme/theme.js
@@ -1,4 +1,3 @@
 "use strict";
 
 /*@*/require( '../docviewjs/theme/default/theme.js' );
-/*@*/require( './google-code-prettify/prettify.js' );